Refactor away from Hook This was to avoid re-renders triggering fetch
